refactor: move emotion helpers to IMS service
