+++ zihuan_core/src/agent/utils/mod.rs
@@ -1,26 +1,1 @@
 pub mod string_utils;
-
-use crate::agent::emotion::utils::emotion_expression_prompt;
-use crate::agent::session_state::QqChatAgentServiceSessionState;
-use crate::agent::qq_chat::QqChatEmotionDimensionConfig;
-
-/// Builds the prefix lines shared by all user-message construction paths:
-/// `[Agent State Snapshot]` + `[System Instructions]`.
-pub fn build_state_system_prefix_lines(
-    session_state: &QqChatAgentServiceSessionState,
-    emotion_dimensions: &[QqChatEmotionDimensionConfig],
-    character_instructions: &str,
-    preprompt_context: Option<&str>,
-) -> Vec<String> {
-    let mut lines = vec!["**Your character's current state**:".to_string()];
-    let emotion_prompt = emotion_expression_prompt(session_state, emotion_dimensions);
-    if !emotion_prompt.is_empty() {
-        // lines.push(format!("- Your emotion expression instructions: {emotion_prompt}"));
-        lines.push(format!("- {emotion_prompt}"));
-    }
-    lines.push(format!("- Your character instructions: {character_instructions}"));
-    if let Some(block) = preprompt_context.map(str::trim).filter(|text| !text.is_empty()) {
-        lines.push(format!("- [Preprompt Context]\n{block}"));
-    }
-    lines
-}
