+++ zihuan_ims_service/src/agent/mod.rs
@@ -1,1 +1,3 @@
 pub(crate) mod before_brain_agent;
+pub mod emotion;
+pub mod utils;
